Expand useful scale resolution down to the natural width of the double (48 bit mantissa) as opposed to the 32 bit limitations of previous scale setup due to 32 bit integer ops.
Eliminate lockup possibilities from scale underflow; report inability to zoom past possible limits (rendering will get blocky well before the scale calculation udnderflows now).

+++ sushivision/scale.h
@@ -23,7 +23,7 @@ typedef struct {
   double lo;
   double hi;
 
-  long first_val;
+  long long first_val;
   long first_pixel;
   
   long step_val;
